Moved fetchXXXQueryParam() methods to RouteManager.ts
